Add histogram for RGB u8 images
