+++ crates/viewer/re_data_ui/src/image.rs
@@ -1,5 +1,4 @@
 use egui::{Color32, NumExt as _, Vec2};
-use itertools::Itertools as _;
 
 use re_renderer::renderer::ColormappedTexture;
 use re_types::{
@@ -385,114 +384,5 @@ fn image_pixel_value_ui(
         ui.label(text);
     } else {
         ui.label("No Value");
-    }
-}
-
-#[allow(dead_code)] // TODO(#6891): use again when we can view image archetypes in the selection view
-fn rgb8_histogram_ui(ui: &mut egui::Ui, rgb: &[u8]) -> egui::Response {
-    re_tracing::profile_function!();
-
-    let mut histograms = [[0_u64; 256]; 3];
-    {
-        // TODO(emilk): this is slow, so cache the results!
-        re_tracing::profile_scope!("build");
-        for pixel in rgb.chunks_exact(3) {
-            for c in 0..3 {
-                histograms[c][pixel[c] as usize] += 1;
-            }
-        }
-    }
-
-    use egui_plot::{Bar, BarChart, Legend, Plot};
-
-    let names = ["R", "G", "B"];
-    let colors = [Color32::RED, Color32::GREEN, Color32::BLUE];
-
-    let charts = histograms
-        .into_iter()
-        .enumerate()
-        .map(|(component, histogram)| {
-            let fill = colors[component].linear_multiply(0.5);
-
-            BarChart::new(
-                histogram
-                    .into_iter()
-                    .enumerate()
-                    .map(|(i, count)| {
-                        Bar::new(i as _, count as _)
-                            .width(0.9)
-                            .fill(fill)
-                            .vertical()
-                            .stroke(egui::Stroke::NONE)
-                    })
-                    .collect(),
-            )
-            .color(colors[component])
-            .name(names[component])
-        })
-        .collect_vec();
-
-    re_tracing::profile_scope!("show");
-    Plot::new("rgb_histogram")
-        .legend(Legend::default())
-        .height(200.0)
-        .show_axes([false; 2])
-        .show(ui, |plot_ui| {
-            for chart in charts {
-                plot_ui.bar_chart(chart);
-            }
-        })
-        .response
-}
-
-#[allow(dead_code)] // TODO(#6891): use again when we can view image archetypes in the selection view
-#[cfg(not(target_arch = "wasm32"))]
-fn copy_and_save_image_ui(ui: &mut egui::Ui, tensor: &re_types::datatypes::TensorData) {
-    ui.horizontal(|ui| {
-        if tensor.could_be_dynamic_image() && ui.button("Click to copy image").clicked() {
-            match tensor.to_dynamic_image() {
-                Ok(dynamic_image) => {
-                    let rgba = dynamic_image.to_rgba8();
-                    re_viewer_context::Clipboard::with(|clipboard| {
-                        clipboard.set_image(
-                            [rgba.width() as _, rgba.height() as _],
-                            bytemuck::cast_slice(rgba.as_raw()),
-                        );
-                    });
-                }
-                Err(err) => {
-                    re_log::error!("Failed to convert tensor to image: {err}");
-                }
-            }
-        }
-
-        if ui.button("Save image…").clicked() {
-            match tensor.to_dynamic_image() {
-                Ok(dynamic_image) => {
-                    save_image(&dynamic_image);
-                }
-                Err(err) => {
-                    re_log::error!("Failed to convert tensor to image: {err}");
-                }
-            }
-        }
-    });
-}
-
-#[allow(dead_code)] // TODO(#6891): use again when we can view image archetypes in the selection view
-#[cfg(not(target_arch = "wasm32"))]
-fn save_image(dynamic_image: &image::DynamicImage) {
-    if let Some(path) = rfd::FileDialog::new()
-        .set_file_name("image.png")
-        .save_file()
-    {
-        match dynamic_image.save(&path) {
-            Ok(()) => {
-                re_log::info!("Image saved to {path:?}");
-            }
-            Err(err) => {
-                re_log::error!("Failed saving image to {path:?}: {err}");
-            }
-        }
     }
 }
